This is a fix for the Turbolinks blinking problem.
I describe my fix here: https://github.com/jeffreyguenther/vue-turbolinks/issues/16#issuecomment-451629141

This took at least an hour and a half to figure out and resolve. Anyway,
the point is that this commit exists as a test. I'm not sure if I'll go
with this fix, but it does fix the problem, so I'm keeping it here for
when I inevitably fail to get a different working and have to revert.