Backport some changes from the release dates branch.
Add error display to the games form and modify the bulma notification
styling to accommodate for that change. Add a cancel button to the games
form and fix the status the games#update endpoint returns for JSON. Also
pass absolute URLs to the game-form component and add an optional
'required' property to the text-field component.