At least it compiles, I guess.
This adds a graphiql:generate task for generating the react and graphiql
packages that get vendored in the repo. There are a bunch of things I
had to bang my head against to get this to even compile, and now that
it compiles I just get an error about an undefined variable. >.>