Fix a weird issue with wikidata_id validation on various models.
For some reason — due to a change introduced by Rails 6.1 – the backend
was receiving a blank string for the Wikidata ID when nothing was input
into the field, and it didn't like that (because of the numericality
validation).

I'm not sure if this is a correction of bad behavior in older versions
of Rails or if this is a bug, but regardless the fix is relatively
simple so I just work around it.

`numericality` doesn't actually support an `allow_nil` argument as far
as I can tell, so I've removed it and used allow_blank outside the
numericality options hash instead.