Disable client_credentials and enable implicit OAuth flow.
I've disabled client_credentials because it's
presumably not used by anyone and the purpose it serves doesn't really
make sense given that the API Token flow is effectively the same.

Client Credentials are only used for machine-to-machine authentication,
and were really only supported by accident.

The implicit flow can only have the read scope because it's not as
secure. It's intended to be used for SPAs and in other places
where an application secret cannot be used.