Create a Rake task for generating variant records.
This enables track_variants for the Rails app.

This also fixes an issue with the return type of sized_avatar and
sized_cover when track_variants is set to true.

The Rake task will generate a processed variant for every cover and
avatar we have, in all the sizes we use. This will cause ActiveStorage
to save the record for the variant in the database. Thus, we avoid the
problem of users generating dozens or hundreds of records when they
use the site. For new images uploaded to the site, the problem
will still exist, but it shouldn't be a big enough deal to really
matter as most common variants will be generated soon after upload by
the user viewing the resulting page.

This Rake task will upload new files in your storage service if you
don't already have variants generated for any covers or avatars on the
site, which is a side-effect worth being aware of. But it's not really
a big problem in this case.