Update MobyGames IDs to the new format.
MobyGames released "v2" a few months ago and switched their IDs from
slugs to ints.

This will delete all the existing MobyGames IDs in the database, but
they all came from Wikidata anyway, so we can just import the updated
MobyGames IDs after this is deployed.

This is a breaking change to the GraphQL API, but I don't think it's a
big enough one that we should go to the trouble of a deprecation cycle.